requirements.txt
jinja2
pandas
requests
python-dotenv